+++ app/src/main/java/com/reintegro/profuturo/app/data/entity/ClientEntity.java
@@ -46,7 +46,7 @@ public class ClientEntity {
     private boolean phoneValid;
     private boolean selected;
     private long agreementId;
-    private long biometricIndicatorValue;
+    private int biometricIndicatorValue;
     private long clientTypeIndicatorValue;
     private long identificationIndicatorValue;
     private long imageId;
@@ -397,11 +397,11 @@ public class ClientEntity {
         this.agreementId = agreementId;
     }
 
-    public long getBiometricIndicatorValue() {
+    public int getBiometricIndicatorValue() {
         return biometricIndicatorValue;
     }
 
-    public void setBiometricIndicatorValue(long biometricIndicatorValue) {
+    public void setBiometricIndicatorValue(int biometricIndicatorValue) {
         this.biometricIndicatorValue = biometricIndicatorValue;
     }
